v1/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML==6.0
regex==2023.3.23
requests==2.28.2
